- fix package and module name references in README's - fix the `dispatch-android-viewmodel` module's base package   - closes #114
