+++ dispatch-android-viewmodel/src/test/java/dispatch/android/viewmodel/ViewModelScopeFactoryTest.kt
@@ -13,9 +13,8 @@
  * limitations under the License.
  */
 
-package dispatch.android.lifecycle
+package dispatch.android.viewmodel
 
-import androidx.lifecycle.*
 import dispatch.core.*
 import dispatch.internal.test.*
 import io.kotest.matchers.*
@@ -27,7 +26,7 @@ import kotlin.coroutines.*
 
 @ObsoleteCoroutinesApi
 @ExperimentalCoroutinesApi
-internal class CoroutineViewModelTest {
+internal class ViewModelScopeFactoryTest {
 
   val job = Job()
   val dispatcher = newSingleThreadContext("single thread dispatcher")
@@ -55,37 +54,15 @@ internal class CoroutineViewModelTest {
   }
 
   @Test
-  fun `accessing the property repeatedly should return the same scope`() = runBlocking {
+  fun `default factory should be a default MainCoroutineScope`() = runBlockingTest {
 
-    val viewModel = TestViewModel()
-
-    val fresh = viewModel.viewModelScope
-
-    val cached = viewModel.viewModelScope
-
-    fresh shouldBeSameInstanceAs cached
-  }
-
-  @Test
-  fun `different ViewModels should get different scopes`() = runBlocking {
-
-    val fresh = TestViewModel().viewModelScope
-
-    val alsoFresh = TestViewModel().viewModelScope
-
-    fresh shouldNotBe alsoFresh
-  }
-
-  @Test
-  fun `default factory should be a default MainImmediateCoroutineScope`() = runBlockingTest {
-
-    val scope = TestViewModel().viewModelScope
+    val scope = ViewModelScopeFactory.create()
 
     scope.coroutineContext[DispatcherProvider]!!.shouldBeTypeOf<DefaultDispatcherProvider>()
 
     scope.coroutineContext[Job]!!.shouldBeSupervisorJob()
 
-    scope.coroutineContext[ContinuationInterceptor] shouldBe Dispatchers.Main.immediate
+    scope.coroutineContext[ContinuationInterceptor] shouldBe Dispatchers.Main
 
     scope.shouldBeInstanceOf<MainCoroutineScope>()
   }
@@ -93,31 +70,33 @@ internal class CoroutineViewModelTest {
   @Test
   fun `a custom factory should be used after being set`() = runBlockingTest {
 
-    ViewModelScopeFactory.set { MainImmediateCoroutineScope(originContext) }
+    ViewModelScopeFactory.set { MainCoroutineScope(originContext) }
 
-    val scope = TestViewModel().viewModelScope
+    val scope = ViewModelScopeFactory.create()
 
     scope.coroutineContext shouldEqualFolded originContext + mainDispatcher
   }
 
   @Test
-  fun `an initialized scope should be cancelled during onCleared`() = runBlockingTest {
+  fun `reset after setting a custom factory should return to the default`() = runBlockingTest {
 
-    val store = ViewModelStore()
+    ViewModelScopeFactory.set { MainCoroutineScope(originContext) }
 
-    val owner = ViewModelStoreOwner { store }
+    val custom = ViewModelScopeFactory.create()
 
-    val viewModel = ViewModelProvider(owner).get(TestViewModel::class.java)
+    custom.coroutineContext shouldEqualFolded originContext + mainDispatcher
 
-    val scope = viewModel.viewModelScope
+    ViewModelScopeFactory.reset()
 
-    scope.isActive shouldBe true
+    val default = ViewModelScopeFactory.create()
 
-    store.clear()
+    default.coroutineContext[DispatcherProvider]!!.shouldBeTypeOf<DefaultDispatcherProvider>()
 
-    scope.isActive shouldBe false
+    default.coroutineContext[Job]!!.shouldBeSupervisorJob()
+
+    default.coroutineContext[ContinuationInterceptor] shouldBe Dispatchers.Main
+
+    default.shouldBeInstanceOf<MainCoroutineScope>()
   }
-
-  class TestViewModel : CoroutineViewModel()
 
 }
