+++ docs/dokka/dispatch-android-viewmodel/alltypes/index.md
@@ -8,15 +8,15 @@ The artifact I hope you don't need, but if you're not doing dependency injection
 |---|---|
 |
 
-##### [dispatch.android.lifecycle.CoroutineViewModel](../dispatch.android.lifecycle/-coroutine-view-model/index.md)
+##### [dispatch.android.viewmodel.CoroutineViewModel](../dispatch.android.viewmodel/-coroutine-view-model/index.md)
 
-Base class for [ViewModel](https://developer.android.com/reference/androidx/androidx/lifecycle/ViewModel.html)s which will be using a [viewModelScope](../dispatch.android.lifecycle/-coroutine-view-model/view-model-scope.md).
+Base class for [ViewModel](https://developer.android.com/reference/androidx/androidx/lifecycle/ViewModel.html)s which will be using a [viewModelScope](../dispatch.android.viewmodel/-coroutine-view-model/view-model-scope.md).
 
 
 |
 
-##### [dispatch.android.lifecycle.ViewModelScopeFactory](../dispatch.android.lifecycle/-view-model-scope-factory/index.md)
+##### [dispatch.android.viewmodel.ViewModelScopeFactory](../dispatch.android.viewmodel/-view-model-scope-factory/index.md)
 
-Factory holder for [viewModelScope](../dispatch.android.lifecycle/-coroutine-view-model/view-model-scope.md)'s.
+Factory holder for [viewModelScope](#)'s.
