--- docs/dokka/dispatch-android-viewmodel/dispatch.android.lifecycle/-coroutine-view-model/on-clear.md
@@ -1,8 +1,0 @@
-[dispatch-android-viewmodel](../../index.md) / [dispatch.android.lifecycle](../index.md) / [CoroutineViewModel](index.md) / [onClear](./on-clear.md)
-
-# onClear
-
-`protected open fun onClear(): `[`Unit`](https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-unit/index.html) [(source)](https://github.com/RBusarow/Dispatch/tree/master/dispatch-android-viewmodel/src/main/java/dispatch/android/lifecycle/CoroutineViewModel.kt#L85)
-
-This method will be called when this ViewModel is no longer used and will be destroyed.
-
