+++ docs/dokka/dispatch-android-viewmodel/dispatch.android.viewmodel/-view-model-scope-factory/index.md
@@ -1,10 +1,10 @@
-[dispatch-android-viewmodel](../../index.md) / [dispatch.android.lifecycle](../index.md) / [ViewModelScopeFactory](./index.md)
+[dispatch-android-viewmodel](../../index.md) / [dispatch.android.viewmodel](../index.md) / [ViewModelScopeFactory](./index.md)
 
 # ViewModelScopeFactory
 
-`object ViewModelScopeFactory` [(source)](https://github.com/RBusarow/Dispatch/tree/master/dispatch-android-viewmodel/src/main/java/dispatch/android/lifecycle/ViewModelScopeFactory.kt#L35)
+`object ViewModelScopeFactory` [(source)](https://github.com/RBusarow/Dispatch/tree/master/dispatch-android-viewmodel/src/main/java/dispatch/android/viewmodel/ViewModelScopeFactory.kt#L35)
 
-Factory holder for [viewModelScope](../-coroutine-view-model/view-model-scope.md)'s.
+Factory holder for [viewModelScope](#)'s.
 
 By default, `create` returns a [MainImmediateCoroutineScope](https://rbusarow.github.io/Dispatch/dispatch-core/dispatch.core/-main-immediate-coroutine-scope/index.md), but may return any [CoroutineScope](https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-coroutine-scope/index.html).
