+++ docs/dokka/dispatch-android-viewmodel/dispatch.android.viewmodel/index.md
@@ -1,10 +1,10 @@
-[dispatch-android-viewmodel](../index.md) / [dispatch.android.lifecycle](./index.md)
+[dispatch-android-viewmodel](../index.md) / [dispatch.android.viewmodel](./index.md)
 
-## Package dispatch.android.lifecycle
+## Package dispatch.android.viewmodel
 
 ### Types
 
 | Name | Summary |
 |---|---|
 | [CoroutineViewModel](-coroutine-view-model/index.md) | Base class for [ViewModel](https://developer.android.com/reference/androidx/androidx/lifecycle/ViewModel.html)s which will be using a [viewModelScope](-coroutine-view-model/view-model-scope.md).`abstract class CoroutineViewModel : `[`ViewModel`](https://developer.android.com/reference/androidx/androidx/lifecycle/ViewModel.html) |
-| [ViewModelScopeFactory](-view-model-scope-factory/index.md) | Factory holder for [viewModelScope](-coroutine-view-model/view-model-scope.md)'s.`object ViewModelScopeFactory` |
+| [ViewModelScopeFactory](-view-model-scope-factory/index.md) | Factory holder for [viewModelScope](#)'s.`object ViewModelScopeFactory` |
