+++ docs/dokka/dispatch-android-viewmodel/package-list
@@ -1,4 +1,4 @@
 $dokka.format:gfm
 $dokka.linkExtension:md
 
-dispatch.android.lifecycle
+dispatch.android.viewmodel
