+++ docs/modules/dispatch-detekt.md
@@ -117,14 +117,14 @@ There is an [issue](https://github.com/detekt/detekt/issues/2582) with ClassLoad
 [Deferred.onAwait]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-deferred/on-await.html
 [Deferred]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-deferred/index.html
 [delay]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/delay.html
-[dispatch-android-espresso]: https://rbusarow.github.io/Dispatch/android-espresso//index.html
-[dispatch-android-lifecycle-extensions]: https://rbusarow.github.io/Dispatch/android-lifecycle-extensions//index.html
-[dispatch-android-lifecycle]: https://rbusarow.github.io/Dispatch/android-lifecycle//index.html
-[dispatch-android-viewmodel]: https://rbusarow.github.io/Dispatch/android-lifecycle-viewmodel//index.html
-[dispatch-core-test-junit4]: https://rbusarow.github.io/Dispatch/core-test-junit4//index.html
-[dispatch-core-test-junit5]: https://rbusarow.github.io/Dispatch/core-test-junit5//index.html
-[dispatch-core-test]: https://rbusarow.github.io/Dispatch/core-test//index.html
-[dispatch-core]: https://rbusarow.github.io/Dispatch/core//index.html
+[dispatch-android-espresso]: https://rbusarow.github.io/Dispatch/dispatch-android-espresso//index.html
+[dispatch-android-lifecycle-extensions]: https://rbusarow.github.io/Dispatch/dispatch-android-lifecycle-extensions//index.html
+[dispatch-android-lifecycle]: https://rbusarow.github.io/Dispatch/dispatch-android-lifecycle//index.html
+[dispatch-android-viewmodel]: https://rbusarow.github.io/Dispatch/dispatch-android-lifecycle-viewmodel//index.html
+[dispatch-test-junit4]: https://rbusarow.github.io/Dispatch/dispatch-test-junit4//index.html
+[dispatch-test-junit5]: https://rbusarow.github.io/Dispatch/dispatch-test-junit5//index.html
+[dispatch-dispatch-test]: https://rbusarow.github.io/Dispatch/dispatch-test//index.html
+[dispatch-core]: https://rbusarow.github.io/Dispatch/dispatch-core//index.html
 [dispatch-extensions]: https://rbusarow.github.io/Dispatch/extensions//index.html
 [Dispatchers.Default]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-dispatchers/-default.html
 [Dispatchers.IO]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-dispatchers/-io.html
