+++ docs/modules/dispatch-test-junit4.md
@@ -5,7 +5,7 @@
 
 * [Features](#features)
 * [Setting Dispatchers.Main](#setting-dispatchersmain)
-* [This module replaces core-test](#this-module-replaces-core-test)
+* [This module replaces dispatch-test](#this-module-replaces-dispatch-test)
 * [JUnit dependencies](#junit-dependencies)
   * [Minimum Gradle Config](#minimum-gradle-config)
   * [JUnit 5 interoperability](#junit-5-interoperability)
@@ -14,7 +14,7 @@
 
 ## Features
 
-In addition to all the functionality in [dispatch-core-test], this module exposes a [TestCoroutineRule] to handle set-up and tear-down of a [TestProvidedCoroutineScope].
+In addition to all the functionality in [dispatch-dispatch-test], this module exposes a [TestCoroutineRule] to handle set-up and tear-down of a [TestProvidedCoroutineScope].
 
 Since [TestProvidedCoroutineScope] is a [TestCoroutineScope], this Rule also invokes [cleanupTestCoroutines][cleanupTestCoroutines] after the test.
 
@@ -46,9 +46,9 @@ class SomeClass(val coroutineScope: CoroutineScope) {
 
 Even though `dispatch-core` eliminates the need to use `Dispatchers.Main` in internal code, it’s still possible that code which has yet to be migrated, or a third-party library is making use of the hard-coded dispatcher.  Because of this, the rule still calls `Dispatchers.setMain(...)` in its setup and `Dispatchers.resetMain()` afterwards.
 
-## This module replaces core-test
+## This module replaces dispatch-test
 
-If using this module, there is no need to include the `dispatch-core-test` artifact in your dependencies.
+If using this module, there is no need to include the `dispatch-dispatch-test` artifact in your dependencies.
 
 ## JUnit dependencies
 
@@ -78,8 +78,8 @@ dependencies {
   implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7"
   implementation "com.rickbusarow.dispatch:dispatch-core:1.0.0-beta03"
 
-  // the junit4 artifact also provides the dispatch-core-test artifact
-  testImplementation "com.rickbusarow.dispatch:dispatch-core-test-junit4:1.0.0-beta03"
+  // the junit4 artifact also provides the dispatch-dispatch-test artifact
+  testImplementation "com.rickbusarow.dispatch:dispatch-test-junit4:1.0.0-beta03"
   testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.7"
 }
 ```
@@ -103,8 +103,8 @@ dependencies {
   implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7")
   implementation("com.rickbusarow.dispatch:dispatch-core:1.0.0-beta03")
 
-  // the junit4 artifact also provides the dispatch-core-test artifact
-  testImplementation("com.rickbusarow.dispatch:dispatch-core-test-junit4:1.0.0-beta03")
+  // the junit4 artifact also provides the dispatch-dispatch-test artifact
+  testImplementation("com.rickbusarow.dispatch:dispatch-test-junit4:1.0.0-beta03")
   testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.7")
 }
 ```
@@ -131,9 +131,9 @@ dependencies {
   implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7"
   implementation "com.rickbusarow.dispatch:dispatch-core:1.0.0-beta03"
 
-  // the junit4 and junit5 artifacts also provides the dispatch-core-test artifact
-  testImplementation "com.rickbusarow.dispatch:dispatch--core-test-junit4:1.0.0-beta03"
-  testImplementation "com.rickbusarow.dispatch:dispatch--core-test-junit5:1.0.0-beta03"
+  // the junit4 and junit5 artifacts also provides the dispatch-dispatch-test artifact
+  testImplementation "com.rickbusarow.dispatch:dispatch--dispatch-test-junit4:1.0.0-beta03"
+  testImplementation "com.rickbusarow.dispatch:dispatch--dispatch-test-junit5:1.0.0-beta03"
   testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.7"
 }
 ```
@@ -155,9 +155,9 @@ dependencies {
   implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.7")
   implementation("com.rickbusarow.dispatch:dispatch-core:1.0.0-beta03")
 
-  // the junit4 and junit5 artifacts also provides the dispatch-core-test artifact
-  testImplementation("ccom.rickbusarow.dispatch:dispatch--core-test-junit4:1.0.0-beta03")
-  testImplementation("ccom.rickbusarow.dispatch:dispatch--core-test-junit5:1.0.0-beta03")
+  // the junit4 and junit5 artifacts also provides the dispatch-dispatch-test artifact
+  testImplementation("ccom.rickbusarow.dispatch:dispatch--dispatch-test-junit4:1.0.0-beta03")
+  testImplementation("ccom.rickbusarow.dispatch:dispatch--dispatch-test-junit5:1.0.0-beta03")
   testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.7")
 }
 ```
@@ -174,6 +174,6 @@ dependencies {
 
 [cleanupTestCoroutines]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-test/kotlinx.coroutines.test/-test-coroutine-scope/cleanup-test-coroutines.html
 
-[dispatch-core-test]: https://rbusarow.github.io/Dispatch/core-test//index.html
+[dispatch-dispatch-test]: https://rbusarow.github.io/Dispatch/dispatch-test//index.html
 [TestCoroutineScope]: https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-test/kotlinx.coroutines.test/-test-coroutine-scope/index.html
